Move addon manager to common and use in xterm-core
