+++ src/common/public/AddonManager.test.ts
