+++ src/common/public/AddonManager.ts
